+++ src/main/java/net/sf/mpxj/CostRateTable.java
@@ -78,5 +78,4 @@ public class CostRateTable extends ArrayList<CostRateTableEntry>
 
       return result;
    }
-
 }