+++ src/main/java/net/sf/mpxj/mspdi/DatatypeConverter.java
@@ -1151,7 +1151,18 @@ public final class DatatypeConverter
     */
    public static final BigDecimal printCurrency(Number value)
    {
-      return (value == null || value.doubleValue() == 0 ? null : new BigDecimal(value.doubleValue() * 100));
+      return value == null || value.doubleValue() == 0 ? null : printCurrencyMandatory(value);
+   }
+
+   /**
+    * Print currency.
+    * 
+    * @param value currency value
+    * @return currency value
+    */
+   public static final BigDecimal printCurrencyMandatory(Number value)
+   {
+      return value == null || value.doubleValue() == 0 ? new BigDecimal(0) : new BigDecimal(value.doubleValue() * 100);
    }
 
    /**
@@ -1665,12 +1676,18 @@ public final class DatatypeConverter
     */
    public static final BigDecimal printRate(Rate rate)
    {
-      BigDecimal result = null;
-      if (rate != null && rate.getAmount() != 0)
-      {
-         result = new BigDecimal(rate.getAmount());
-      }
-      return result;
+      return rate == null || rate.getAmount() == 0 ? null : printRateMandatory(rate);
+   }
+
+   /**
+    * Print rate.
+    *
+    * @param rate Rate instance
+    * @return rate value
+    */
+   public static final BigDecimal printRateMandatory(Rate rate)
+   {
+      return rate == null || rate.getAmount() == 0 ? new BigDecimal(0) : new BigDecimal(rate.getAmount());
    }
 
    /**